connmgr: Tighten TestRemovePendingConnection.
This reworks the TestRemovePendingConnection test to more accurately
test the intended condition without dialer sleep hacks.

It does this by signaling exactly when a dial attempt has been made and
responding to that signal by removing the the pending connection while
the dialer is intentionally blocking and ensuring the connection request
is marked as canceled.

It also adds a timeout to the initial dial to prevent a test timeout in
the case the connection were to fail.