Break conext-free functions apart.
Due to the introduction of the treasury agenda both checkBlockSanity and
checkTransactionSanity have become contextual. This is obviously
incorret however a proper fix will require quite a bit of shuffling of
code. Davec and I have agreed upon a temporary hack where the context
free and contextual pieces are two different functions. This shall be
cleaned up later.